// empleados_cedenet/interfaces/users.tsx
export type DataUpdateUser = {
    ID?: number
    primer_nombre?: string
    primer_apellido?: string   
    segundo_apellido?: string
    otros_nombres?: string
    pais?: string
    tipo_documento?: string 
    numero_identificacion?: number
    email?: string
    fecha_ingreso?: Date 
    area?: string
    estado?: string
}